Moved code from __init__ to appknox
